Don't run migration operations in ReTypers
